webapp code rearrangement
