+++ src/app.py
@@ -8,38 +8,15 @@ import logging
 import os
 import re
 import sys
-import time
-from typing import Dict, Set
 import urllib.parse
 
-from webapp import contacts_reader, project_reader, rg_reader, vo_reader, default_config
-from webapp.common import git_clone_or_pull, to_xml_bytes, Filters
-from webapp.contacts_reader import ContactsData
-from webapp.topology import GRIDTYPE_1, GRIDTYPE_2, Topology
-from webapp.vos_data import VOsData
+from webapp import default_config
+from webapp.common import to_xml_bytes, Filters
+from webapp.models import GlobalData
+from webapp.topology import GRIDTYPE_1, GRIDTYPE_2
 
 
 class InvalidArgumentsError(Exception): pass
-
-class DataError(Exception): pass
-
-
-class CachedData:
-    def __init__(self, data=None, timestamp=0, force_update=True, cache_lifetime=60*15):
-        self.data = data
-        self.timestamp = timestamp
-        self.force_update = force_update
-        self.cache_lifetime = cache_lifetime
-
-    def should_update(self):
-        return self.force_update or (not self.data or time.time() - self.timestamp > self.cache_lifetime)
-
-    def update(self, data, bump_timestamp=True):
-        self.data = data
-        if bump_timestamp:
-            self.timestamp = time.time()
-        self.force_update = False
-
 
 def _verify_config(cfg):
     if not cfg["NO_GIT"]:
@@ -52,94 +29,6 @@ def _verify_config(cfg):
             st = os.stat(ssh_key)
             if st.st_uid != os.getuid() or (st.st_mode & 0o7777) not in (0o700, 0o600, 0o400):
                 raise PermissionError(ssh_key)
-
-
-class GlobalData:
-    def __init__(self, config):
-        self.contacts_data = CachedData(cache_lifetime=config["CACHE_LIFETIME"])
-        self.dn_set = CachedData(cache_lifetime=config["CACHE_LIFETIME"])
-        self.projects = CachedData(cache_lifetime=config["CACHE_LIFETIME"])
-        self.topology = CachedData(cache_lifetime=config["CACHE_LIFETIME"])
-        self.vos_data = CachedData(cache_lifetime=config["CACHE_LIFETIME"])
-        self.contacts_file = os.path.join(config["CONTACT_DATA_DIR"], "contacts.yaml")
-        self.projects_dir = os.path.join(config["TOPOLOGY_DATA_DIR"], "projects")
-        self.topology_dir = os.path.join(config["TOPOLOGY_DATA_DIR"], "topology")
-        self.vos_dir = os.path.join(config["TOPOLOGY_DATA_DIR"], "virtual-organizations")
-        self.config = config
-
-    def _update_topology_repo(self):
-        if not self.config["NO_GIT"]:
-            parent = os.path.dirname(self.config["TOPOLOGY_DATA_DIR"])
-            os.makedirs(parent, mode=0o755, exist_ok=True)
-            ok = git_clone_or_pull(self.config["TOPOLOGY_DATA_REPO"], self.config["TOPOLOGY_DATA_DIR"],
-                                   self.config["TOPOLOGY_DATA_BRANCH"])
-            if ok:
-                app.logger.debug("topology repo update ok")
-            else:
-                app.logger.warning("topology repo update failed")
-                return False
-        for d in [self.projects_dir, self.topology_dir, self.vos_dir]:
-            if not os.path.exists(d):
-                app.logger.error("%s not in topology repo", d)
-                return False
-        return True
-
-    def _update_contacts_repo(self):
-        if not self.config["NO_GIT"]:
-            parent = os.path.dirname(self.config["CONTACT_DATA_DIR"])
-            os.makedirs(parent, mode=0o700, exist_ok=True)
-            ok = git_clone_or_pull(self.config["CONTACT_DATA_REPO"], self.config["CONTACT_DATA_DIR"],
-                                   self.config["CONTACT_DATA_BRANCH"], self.config["GIT_SSH_KEY"])
-            if ok:
-                app.logger.debug("contact repo update ok")
-            else:
-                app.logger.warning("contact repo update failed")
-                return False
-        if not os.path.exists(self.contacts_file):
-            app.logger.error("%s not in contact repo", self.contacts_file)
-            return False
-        return True
-
-    def get_contacts_data(self) -> ContactsData:
-        """
-        Get the contact information from a private git repo
-        """
-        if self.contacts_data.should_update():
-            ok = self._update_contacts_repo()
-            self.contacts_data.update(contacts_reader.get_contacts_data(self.contacts_file),
-                                      bump_timestamp=ok)
-
-        return self.contacts_data.data
-
-    def get_dns(self) -> Set:
-        """
-        Get the set of DNs allowed to access "special" data (such as contact info)
-        """
-        if self.dn_set.should_update():
-            contacts_data = self.get_contacts_data()
-            self.dn_set.update(set(contacts_data.get_dns()))
-        return self.dn_set.data
-
-    def get_topology(self) -> Topology:
-        if self.topology.should_update():
-            ok = self._update_topology_repo()
-            self.topology.update(rg_reader.get_topology(self.topology_dir, self.get_contacts_data()),
-                                 bump_timestamp=ok)
-        return self.topology.data
-
-    def get_vos_data(self) -> VOsData:
-        if self.vos_data.should_update():
-            ok = self._update_topology_repo()
-            self.vos_data.update(vo_reader.get_vos_data(self.vos_dir, self.get_contacts_data()),
-                                 bump_timestamp=ok)
-        return self.vos_data.data
-
-    def get_projects(self) -> Dict:
-        if self.projects.should_update():
-            ok = self._update_topology_repo()
-            self.projects.update(project_reader.get_projects(self.projects_dir),
-                                 bump_timestamp=ok)
-        return self.projects.data
 
 
 default_authorized = False
